Don't mark non-{js,css} dirs as external
This allows the files inside those other directories to be referenced from within JS/CSS files and properly bundled by esbuild.
